requirements.txt
numpy==1.17.3
Pillow==6.2.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
pandas==0.25.2
opencv-python==4.1.1.26
matplotlib==3.1.1